requirements.txt
watchdog
colorama
pyyaml
flask
requests
cryptography
geoip2
schedule
flask-socketio
eventlet
flask-cors

